Check for newlines and carriage return in artifact paths and name
