+++ packages/artifact/src/internal/upload-specification.ts
@@ -1,7 +1,7 @@
 import * as fs from 'fs'
 import {debug} from '@actions/core'
 import {join, normalize, resolve} from 'path'
-import {checkArtifactFilePath} from './utils'
+import {checkArtifactFilePath} from './path-and-artifact-name-validation'
 
 export interface UploadSpecification {
   absoluteFilePath: string
